feat: Introduce store hydration modes
This commit introduces a positional parameter to the `WithStore` option
which controls how the value returned from the store is used with regard
to the `ReferenceObject`.  There are three "hydration" modes: `always`,
`specnil` ("Spec is nil"), and `never`.  These modes dictate when to set
the value found from the store.  The word "hydration" suggests that the
contents of the value is pre-populated before its use via the Zip API
framework.

Signed-off-by: Alexander Jung <alex@unikraft.io>